requirements.txt
acres==0.5.0
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
camelot-py==1.0.0
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
configobj==5.0.9
configparser==7.2.0
cryptography==45.0.5
cymem==2.0.11
distro==1.9.0
et_xmlfile==2.0.0
etelemetry==0.3.1
acres==0.5.0
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
camelot-py==1.0.0
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
configobj==5.0.9
configparser==7.2.0
cryptography==45.0.5
cymem==2.0.11
distro==1.9.0
et_xmlfile==2.0.0
etelemetry==0.3.1
faiss-cpu==1.11.0
filelock==3.18.0
fitz==0.0.1.dev2
Flask==3.1.1
fsspec==2025.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.22.0
huggingface-hub==0.33.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jpype1==1.6.0
langcodes==3.5.0
language_data==1.3.0
looseversion==1.3.0
lxml==6.0.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pathlib==1.0.1
pathlib2==2.3.7.post1
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
playwright==1.53.0
preshed==3.0.10
proto-plus==1.26.1
protobuf==5.29.5
prov==2.1.1
pt_core_news_lg @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_lg-3.8.0/pt_core_news_lg-3.8.0-py3-none-any.whl#sha256=2561c9a72a938d37141e9694e1a36d25061a44ce7e4f3bad2d3fa3bb836191af
puremagic==1.30
py-trello==0.20.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
pyee==13.0.0
Pygments==2.19.2
PyMuPDF==1.26.3
pyparsing==3.2.3
pypdf==5.9.0
PyPDF2==3.0.1
pypdfium2==4.30.1
PySocks==1.7.1
pytesseract==0.3.13
pytest==8.1.1
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
tabulate==0.9.0
tqdm==4.66.2
filelock==3.18.0
fitz==0.0.1.dev2
Flask==3.1.1
fsspec==2025.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.22.0
huggingface-hub==0.33.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jpype1==1.6.0
langcodes==3.5.0
language_data==1.3.0
looseversion==1.3.0
lxml==6.0.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pathlib==1.0.1
pathlib2==2.3.7.post1
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
playwright==1.53.0
preshed==3.0.10
proto-plus==1.26.1
protobuf==5.29.5
prov==2.1.1
pt_core_news_lg @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_lg-3.8.0/pt_core_news_lg-3.8.0-py3-none-any.whl#sha256=2561c9a72a938d37141e9694e1a36d25061a44ce7e4f3bad2d3fa3bb836191af
puremagic==1.30
py-trello==0.20.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
pyee==13.0.0
Pygments==2.19.2
PyMuPDF==1.26.3
pyparsing==3.2.3
pypdf==5.9.0
PyPDF2==3.0.1
pypdfium2==4.30.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
selenium==4.34.0
selenium-stealth==1.0.6
sentence-transformers==5.0.0
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.14.0
tabula-py==2.10.0
tabulate==0.9.0
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
traits==7.0.2
transformers==4.53.2
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
unicode==2.9
Unidecode==1.4.0
uritemplate==4.2.0
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0